# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Federated Learning API",
        "description": "API documentation for the Medical AI Federated Learning platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/doctor/reports": {
            "get": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "List reports authored by this doctor",
                "operationId": "2888953be9b50d19966aab0ef8862e93",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "final"
                            ]
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of reports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReportObject"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Create a draft report for a concluded examination",
                "operationId": "0869a623a6c00c25cff8d572b964f51e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "examination_id",
                                    "prediction_id"
                                ],
                                "properties": {
                                    "examination_id": {
                                        "type": "integer"
                                    },
                                    "prediction_id": {
                                        "type": "integer"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Examination not concluded or report already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctor/reports/{id}": {
            "get": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Show a single report with full context",
                "operationId": "34474f80711a8effe8655a8a65013c21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Update report notes",
                "operationId": "b83164be6a13c4133d83cf7268affb22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Finalized report cannot be edited"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Delete a draft report",
                "operationId": "24112477440dfe5612c9d42abed74bc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report deleted"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Cannot delete finalized report"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctor/reports/{id}/finalize": {
            "post": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Finalize a report (marks it as official — no further edits allowed)",
                "operationId": "27e2df7ab6a7e4633ffa3e0badae592c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report finalized"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Report already finalized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/doctor/reports/{id}/attach": {
            "post": {
                "tags": [
                    "Doctor — Reports"
                ],
                "summary": "Attach a generated PDF file to the report",
                "operationId": "e514a8d8159dd75d537be71ab85282d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report file attached"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReportObject": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "examination_id": {
                        "type": "integer"
                    },
                    "prediction_id": {
                        "type": "integer"
                    },
                    "patient_id": {
                        "type": "integer"
                    },
                    "doctor_id": {
                        "type": "integer"
                    },
                    "organization_id": {
                        "type": "integer"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "file_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "file_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "final"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "HttpOnly cookie issued by Sanctum after a successful `/api/verify-otp` call.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}